requirements.txt
git+https://github.com/shayypy/guilded.py
python-dotenv
prisma
